# [PT2]: Add Static Dispatch Kernel for wrapped_fbgemm_linear_fp16_weight